requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
certifi==2024.8.30
cffi==1.17.1
channels==3.0.4
channels-redis==4.2.0
charset-normalizer==3.3.2
colorama==0.4.6
constantly==23.10.4
contourpy==1.3.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==43.0.1
cycler==0.12.1
daphne==3.0.2
Django==5.1.1
django-cors-headers==4.4.0
django-rest-swagger==2.2.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
filelock==3.16.1
fonttools==4.54.1
fsspec==2024.9.0
gunicorn==23.0.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.7
MarkupSafe==2.1.5
matplotlib==3.9.2
mpmath==1.3.0
msgpack==1.1.0
networkx==3.3
numpy==1.26.4
openapi-codec==1.3.2
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==24.1
pandas==2.2.3
pillow==10.4.0
psutil==6.0.0
psycopg2-binary==2.9.9
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==24.2.1
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
redis==5.1.0
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.0
scipy==1.14.1
seaborn==0.13.2
service-identity==24.1.0
simplejson==3.19.3
six==1.16.0
sqlparse==0.5.1
sympy==1.13.3
torch==2.4.1
torchvision==0.19.1
tqdm==4.66.5
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
ultralytics==8.3.2
ultralytics-thop==2.0.8
uritemplate==4.1.1
urllib3==2.2.3
zope.interface==7.0.3
